consume handled events
git-svn-id: svn+ssh://svn.code.sf.net/p/jedit/svn/jEdit/trunk@3069 6b1eeb88-9816-0410-afa2-b43733a0f04e

+++ org/gjt/sp/jedit/MiscUtilities.java
@@ -407,6 +407,28 @@ public class MiscUtilities
 		String backupPrefix, String backupSuffix,
 		String backupDirectory)
 	{
+		saveBackup(file,backups,backupPrefix,backupSuffix,backupDirectory,0);
+	} //}}}
+
+	//{{{ saveBackup() method
+	/**
+	 * Saves a backup (optionally numbered) of a file.
+	 * @param file A local file
+	 * @param backups The number of backups. Must be >= 1. If > 1, backup
+	 * files will be numbered.
+	 * @param backupPrefix The backup file name prefix
+	 * @param backupSuffix The backup file name suffix
+	 * @param backupDirectory The directory where to save backups; if null,
+	 * they will be saved in the same directory as the file itself.
+	 * @param backupTimeDistance The minimum time in minutes when a backup
+	 * version 1 shall be moved into version 2; if 0, backups are always
+	 * moved.
+	 * @since jEdit 4.2pre5
+	 */
+	public static void saveBackup(File file, int backups,
+		String backupPrefix, String backupSuffix,
+		String backupDirectory, int backupTimeDistance)
+	{
 		if(backupPrefix == null)
 			backupPrefix = "";
 		if(backupSuffix == null)
@@ -419,30 +441,53 @@ public class MiscUtilities
 		{
 			File backupFile = new File(backupDirectory,
 				backupPrefix + name + backupSuffix);
-			backupFile.delete();
-			file.renameTo(backupFile);
+			long modTime = backupFile.lastModified();
+			/* if backup file was created less than
+			 * 'backupTimeDistance' ago, we do not
+			 * create the backup */
+			if(System.currentTimeMillis() - modTime
+				>= backupTimeDistance)
+			{
+				backupFile.delete();
+				file.renameTo(backupFile);
+			}
 		}
 		// If backups > 1, move old ~n~ files, create ~1~ file
 		else
 		{
+			/* delete a backup created using above method */
 			new File(backupDirectory,
 				backupPrefix + name + backupSuffix
 				+ backups + backupSuffix).delete();
 
-			for(int i = backups - 1; i > 0; i--)
-			{
-				File backup = new File(backupDirectory,
-					backupPrefix + name + backupSuffix
-					+ i + backupSuffix);
-
-				backup.renameTo(new File(backupDirectory,
-					backupPrefix + name + backupSuffix
-					+ (i+1) + backupSuffix));
-			}
-
-			file.renameTo(new File(backupDirectory,
+			File firstBackup = new File(backupDirectory,
 				backupPrefix + name + backupSuffix
-				+ "1" + backupSuffix));
+				+ "1" + backupSuffix);
+			long modTime = firstBackup.lastModified();
+			/* if backup file was created less than
+			 * 'backupTimeDistance' ago, we do not
+			 * create the backup */
+			if(System.currentTimeMillis() - modTime
+				>= backupTimeDistance)
+			{
+				for(int i = backups - 1; i > 0; i--)
+				{
+					File backup = new File(backupDirectory,
+						backupPrefix + name
+						+ backupSuffix + i
+						+ backupSuffix);
+
+					backup.renameTo(
+						new File(backupDirectory,
+						backupPrefix + name
+						+ backupSuffix + (i+1)
+						+ backupSuffix));
+				}
+
+				file.renameTo(new File(backupDirectory,
+					backupPrefix + name + backupSuffix
+					+ "1" + backupSuffix));
+			}
 		}
 	} //}}}
